move context stuff to context repository
