+++ components/Contexts/Settings/Settings.tsx
@@ -2,7 +2,7 @@ import { useEffect, useState } from 'react';
 import { Fieldset } from '@mantine/core';
 import ArchivalForm from '@/components/Contexts/Settings/ArchivalForm';
 import RenameForm from '@/components/Contexts/Settings/RenameForm';
-import { useDataSource } from '@/contexts/DataSourceContext';
+import { useContextRepository } from '@/contexts/DataSourceContext';
 
 export default function Settings({
   contextName,
@@ -11,13 +11,13 @@ export default function Settings({
   contextName: string;
   onClose: () => void;
 }) {
-  const dataSource = useDataSource();
+  const contextRepository = useContextRepository();
   const [destinationContexts, setDestinationContexts] = useState<string[]>([]);
 
   useEffect(() => {
     // Get the contexts for the archival radio options
     const fetchContexts = async () => {
-      const contexts = await dataSource.getContexts();
+      const contexts = await contextRepository.getContexts();
       setDestinationContexts(contexts.filter((c) => c !== contextName));
     };
     fetchContexts();
