+++ components/Layout/CommandPalette.tsx
@@ -5,12 +5,12 @@ import { Modal } from '@mantine/core';
 import { useDisclosure } from '@mantine/hooks';
 import { Spotlight, SpotlightActionData, SpotlightActionGroupData } from '@mantine/spotlight';
 import TaskEditor from '@/components/Tasks/TaskEditor';
-import { useDataSource, useTaskRepository } from '@/contexts/DataSourceContext';
+import { useContextRepository, useTaskRepository } from '@/contexts/DataSourceContext';
 import { Task, TaskStatus } from '@/data/documentTypes/Task';
 
 export default function CommandPalette() {
   const router = useRouter();
-  const dataSource = useDataSource();
+  const contextRepository = useContextRepository();
   const taskRepository = useTaskRepository();
   const [contexts, setContexts] = useState<string[]>([]);
   const [openTasks, setOpenTasks] = useState<Task[]>([]);
@@ -72,12 +72,12 @@ export default function CommandPalette() {
 
   // Subscribe to contexts
   useEffect(() => {
-    const unsubscribe = dataSource.subscribeToContexts(setContexts);
+    const unsubscribe = contextRepository.subscribeToContexts(setContexts);
 
     return () => {
       unsubscribe();
     };
-  }, [dataSource]);
+  }, []);
 
   // Subscribe to open tasks
   useEffect(() => {
@@ -91,7 +91,7 @@ export default function CommandPalette() {
     return () => {
       unsubscribe();
     };
-  }, [dataSource]);
+  }, []);
 
   return (
     <>
