+++ components/Layout/ContextLinks.tsx
@@ -2,25 +2,25 @@ import React, { useEffect, useState } from 'react';
 import { useRouter } from 'next/router';
 import { IconManualGearboxFilled } from '@tabler/icons-react';
 import { NavLink } from '@mantine/core';
-import { useDataSource } from '@/contexts/DataSourceContext';
+import { useContextRepository } from '@/contexts/DataSourceContext';
 
 export default function ContextLinks({
   handleNavLinkClick,
 }: {
   handleNavLinkClick: (url: string) => void;
 }) {
-  const dataSource = useDataSource();
+  const contextRepository = useContextRepository();
   const router = useRouter();
   const query = router.query;
   const [contexts, setContexts] = useState<string[]>([]);
 
   useEffect(() => {
-    const unsubscribe = dataSource.subscribeToContexts(setContexts);
+    const unsubscribe = contextRepository.subscribeToContexts(setContexts);
 
     return () => {
       unsubscribe();
     };
-  }, [dataSource]);
+  }, []);
 
   const links = contexts.map((context) => (
     <NavLink
