+++ data/DataSource.ts
@@ -1,4 +1,5 @@
 import PouchDB from 'pouchdb';
+import { ContextRepository } from '@/data/ContextRepository';
 import { DocumentTypes } from '@/data/documentTypes';
 import { Context } from '@/data/documentTypes/Context';
 import { LocalSettings } from '@/data/documentTypes/LocalSettings';
@@ -8,22 +9,17 @@ import { TaskRepository } from '@/data/TaskRepository';
 import { Logger } from '@/helpers/Logger';
 import { MigrationManager } from './migrations/MigrationManager';
 
-export type UnsubscribeFunction = () => void;
-export type ContextSubscriber = (contexts: string[]) => void;
-
 /**
  * @TODO This whole thing needs error handling
- * @TODO Absolutely use pouchdb indices once the dust settles
- * @TODO It feels like the pouch stuff may belong in a separate class
+ * @TODO Absolutely use pouchdb indices once the dust settles\
  */
 export class DataSource {
   protected db: PouchDB.Database<DocumentTypes>;
 
-  private contextChangesFeed?: PouchDB.Core.Changes<Context>;
-  private contextChangeSubscribers = new Set<ContextSubscriber>();
   private migrationManager: MigrationManager;
   private syncHandler: PouchDB.Replication.Sync<{}> | null = null;
   private taskRepository: TaskRepository | null = null;
+  private contextRepository: ContextRepository | null = null;
 
   public onMigrationStatusChange?: (isMigrating: boolean) => void;
 
@@ -54,15 +50,7 @@ export class DataSource {
     Logger.info('Cleaning up DataSource');
 
     await this.getTaskRepository().cleanup();
-
-    // Cancel active change feeds
-    if (this.contextChangesFeed) {
-      this.contextChangesFeed.cancel();
-      this.contextChangesFeed = undefined;
-    }
-
-    // Clear all subscribers
-    this.contextChangeSubscribers.clear();
+    await this.getContextRepository().cleanup();
 
     Logger.info('DataSource cleanup completed');
   }
@@ -76,6 +64,17 @@ export class DataSource {
     }
 
     return this.taskRepository;
+  }
+
+  /**
+   * Create/return the ContextRepository singleton
+   */
+  getContextRepository(): ContextRepository {
+    if (!this.contextRepository) {
+      this.contextRepository = new ContextRepository(this.db);
+    }
+
+    return this.contextRepository;
   }
 
   /**
@@ -289,41 +288,6 @@ export class DataSource {
   }
 
   /**
-   * Fetch all context names from the database.
-   */
-  async getContexts(includeDeleted?: boolean): Promise<string[]> {
-    const _includeDeleted = includeDeleted ?? false;
-
-    const result = await this.db.allDocs<Context>({
-      include_docs: true,
-      startkey: 'context-',
-      endkey: 'context-\ufff0',
-    });
-
-    // Filter out archived contexts if not requested
-    if (!_includeDeleted) {
-      result.rows = result.rows.filter((row) => !row.doc?.deletedAt);
-    }
-
-    return result.rows.map((row) => row.doc!.name);
-  }
-
-  /**
-   * Add a new context to the database.
-   * This will create a new document with the type 'context' and the provided name.
-   *
-   * @param context
-   */
-  async addContext(context: string): Promise<void> {
-    const doc: Context = {
-      _id: `context-${context}`,
-      type: 'context',
-      name: context,
-    };
-    await this.db.put(doc);
-  }
-
-  /**
    * Export all data from the database, excluding _local documents.
    */
   async exportAllData(): Promise<DocumentTypes[]> {
@@ -364,39 +328,6 @@ export class DataSource {
     }
   }
 
-  /**
-   * Subscribe to changes in contexts. This will initially fetch all contexts and invoke the
-   * callback with them, then set up a live changes feed to watch for updates, invoking the
-   * callback with the updated contexts whenever a change occurs.
-   *
-   * The return function should be used to unsubscribe from the changes feed when no longer needed
-   * or when the component using this is unmounted.
-   *
-   * @param callback
-   *
-   * @return A function to unsubscribe from the changes feed.
-   */
-  subscribeToContexts(callback: ContextSubscriber): UnsubscribeFunction {
-    // Register the callback so that we can notify it of changes
-    this.contextChangeSubscribers.add(callback);
-
-    // Set up the PouchDB changes feed if this is the first subscriber
-    if (this.contextChangeSubscribers.size === 1) {
-      this.initializeContextChangesFeed();
-    }
-
-    // Provide the initial contexts to the callback
-    try {
-      this.getContexts().then((contexts) => callback(contexts));
-    } catch (error) {
-      // TODO
-      Logger.error('Error fetching initial contexts for watcher:', error);
-    }
-
-    // Return an unsubscribe function
-    return () => this.removeContextSubscriber(callback);
-  }
-
   private async verifySyncConnection(syncDb: PouchDB.Database): Promise<void> {
     try {
       await syncDb.info();
@@ -405,72 +336,6 @@ export class DataSource {
       throw new Error(
         'Failed to connect to sync database. Please check your sync server URL and access token.'
       );
-    }
-  }
-
-  /**
-   * Initialize the PouchDB changes feed to listen for changes to context documents.
-   * This will set up a live feed that listens for any changes to documents with IDs starting
-   * with 'context-' and notifies subscribers of the updated contexts.
-   * @private
-   */
-  private initializeContextChangesFeed(): void {
-    Logger.info('Initializing PouchDB changes feed for contexts');
-    this.contextChangesFeed = this.db
-      .changes<Context>({
-        live: true,
-        since: 'now',
-      })
-      .on('change', async (change) => {
-        // Check if the changed document's ID starts with our context prefix
-        if (change.id.startsWith('context-')) {
-          try {
-            const updatedContexts = await this.getContexts();
-            this.notifyContextSubscribers(updatedContexts);
-          } catch (error) {
-            // TODO
-            Logger.error('Error fetching updated contexts after change:', error);
-          }
-        }
-      })
-      .on('error', (err) => {
-        // TODO
-        Logger.error('Error in PouchDB changes feed for contexts:', err);
-      });
-  }
-
-  /**
-   * Notify all subscribers of context changes.
-   *
-   * @param contexts
-   * @private
-   */
-  private notifyContextSubscribers(contexts: string[]): void {
-    Logger.info('Notifying context change subscribers');
-    this.contextChangeSubscribers.forEach((callback) => {
-      try {
-        callback(contexts);
-      } catch (error) {
-        // TODO
-        Logger.error('Error notifying context change subscriber:', error);
-      }
-    });
-  }
-
-  /**
-   * Remove a context subscriber and cancel the changes feed if there are no more subscribers.
-   *
-   * @param callback
-   * @private
-   */
-  private removeContextSubscriber(callback: ContextSubscriber): void {
-    Logger.info('Removing context change subscriber');
-    this.contextChangeSubscribers.delete(callback);
-
-    if (this.contextChangeSubscribers.size === 0 && this.contextChangesFeed) {
-      Logger.info('No more context subscribers, cancelling changes feed');
-      this.contextChangesFeed.cancel();
-      this.contextChangesFeed = undefined;
     }
   }
