+++ data/Repository.ts
@@ -1,4 +1,4 @@
-interface Repository {
+export interface Repository {
   /**
    * Cleanup all active subscriptions and resources.
    * This should be called when the LocalDataSource instance is no longer needed
